linkedin_agent/requirements.txt
langgraph>=0.2.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=8.0.0
